+++ node/src/test/java/net/corda/node/services/vault/VaultQueryJavaTests.java
@@ -28,8 +28,8 @@ import net.corda.node.services.api.IdentityServiceInternal;
 import net.corda.node.services.persistence.NodeAttachmentService;
 import net.corda.nodeapi.internal.persistence.CordaPersistence;
 import net.corda.nodeapi.internal.persistence.DatabaseTransaction;
-import net.corda.testing.core.ContractJarTestUtils;
-import net.corda.testing.core.SelfCleaningDir;
+import net.corda.testing.core.internal.ContractJarTestUtils;
+import net.corda.testing.core.internal.SelfCleaningDir;
 import net.corda.testing.core.SerializationEnvironmentRule;
 import net.corda.testing.core.TestIdentity;
 import net.corda.testing.internal.TestingNamedCacheFactory;
@@ -58,7 +58,7 @@ import static net.corda.core.node.services.vault.Builder.equal;
 import static net.corda.core.node.services.vault.Builder.sum;
 import static net.corda.core.node.services.vault.QueryCriteriaUtils.*;
 import static net.corda.core.utilities.ByteArrays.toHexString;
-import static net.corda.testing.core.ContractJarTestUtils.INSTANCE;
+import static net.corda.testing.core.internal.ContractJarTestUtils.INSTANCE;
 import static net.corda.testing.core.TestConstants.*;
 import static net.corda.testing.internal.RigorousMockKt.rigorousMock;
 import static net.corda.testing.node.MockServices.makeTestDatabaseAndMockServices;
