+++ testing/test-utils/src/main/kotlin/net/corda/testing/core/internal/ContractJarTestUtils.kt
@@ -1,12 +1,13 @@
-package net.corda.testing.core
+package net.corda.testing.core.internal
 
-import net.corda.testing.core.JarSignatureTestUtils.addManifest
-import net.corda.testing.core.JarSignatureTestUtils.createJar
-import net.corda.testing.core.JarSignatureTestUtils.generateKey
-import net.corda.testing.core.JarSignatureTestUtils.signJar
+import net.corda.testing.core.internal.JarSignatureTestUtils.addManifest
+import net.corda.testing.core.internal.JarSignatureTestUtils.createJar
+import net.corda.testing.core.internal.JarSignatureTestUtils.generateKey
+import net.corda.testing.core.internal.JarSignatureTestUtils.signJar
 import net.corda.core.internal.delete
 import net.corda.core.internal.div
 import net.corda.core.internal.toPath
+import net.corda.testing.core.ALICE_NAME
 import java.io.OutputStream
 import java.net.URI
 import java.net.URL
